Docker build success wih scripts
This currently requires the user has a `$HOME/tmp` folder on the local
system where packer is running. This is a docker volume issue explained
more here
http://chiefy.github.io/2015/02/24/up-and-running-w-boot2docker-fig-packer.html

There needs be a build script created, but for now to run the packer
build properly use something similar to:

```
TMPDIR=~/tmp PACKER_LOG=debug packer build -only=docker -var install_chef_client=true -var-file=var-files/aws.json -var-file=var-files/ubuntu-12.04.json template.json
```

Note: This is not the preferred method of installing chef!!!

+++ scripts/disable_root.sh
@@ -1,12 +1,16 @@
 # Expire the root account.
 disable_root() {
   echo "disable_root() function called"
-  # Remove the root account SSH key.
-  sudo rm -rf /root/.ssh/
-  # Expire the root account password.
-  sudo passwd -dl root
-  # Disable the actual root account
-  usermod --expiredate 1 root
+  if [ "$DISABLE_ROOT" = "true" ]; then
+    # Remove the root account SSH key.
+    sudo rm -rf /root/.ssh/
+    # Expire the root account password.
+    sudo passwd -dl root
+    # Disable the actual root account
+    usermod --expiredate 1 root
+  else
+    echo "Skipping disable_root"
+  fi
 }
 
 disable_root || echo "disable_root() failed"
